add windom confirm as alt way for popup dialog

+++ src/Components/Main.jsx
@@ -15,13 +15,17 @@ const Main = () => {
             // in findAll we set {authors: authors}
             .then(res => setAuthors(res.data.authors))
             .catch(err => console.log(err))
-    },[])
+    }, [])
 
-    const deleteAuthor = (deleteId) => {
-        axios.delete(`http://localhost:8000/api/authors/${deleteId}`)
-            .then(res => setAuthors(authors.filter((author, idx) => author._id !== deleteId)))
-            .catch(err => console.log(err))
+    const deleteAuthor = (deleteId, name) => {
+        if (window.confirm(`Are you sure you want to remove ${name}?`)) {
+            axios.delete(`http://localhost:8000/api/authors/${deleteId}`)
+                .then(res => setAuthors(authors.filter((author, idx) => author._id !== deleteId)))
+                .catch(err => console.log(err))
+        }
     }
+
+
 
     return <div>
         <Link to="/authors/new">Add an author</Link>
@@ -43,10 +47,12 @@ const Main = () => {
                                     <Link to={"/authors/" + author._id}>{author.name}</Link>
                                 </td>
                                 <td>
-                                    <Link className="btn" to={"/authors/" + author._id+"/edit"}>Edit</Link>
+                                    <Link className="btn" to={"/authors/" + author._id + "/edit"}>Edit</Link>
                                     &nbsp;
-                                    <PopUp deleteAuthor={deleteAuthor} author={author}/>
-                                    {/* <button className="btn" onClick={()=>deleteAuthor(author._id)}>Delete</button> */}
+                                    {/* PopUp is another way to do pop up window with custom dialog*/}
+                                    {/* <PopUp deleteAuthor={deleteAuthor} author={author}/> */}
+                                    {/* use windows.confirm to do pop up*/}
+                                    <button className="btn" onClick={() => deleteAuthor(author._id, author.name)}>Delete</button>
                                 </td>
                             </tr>
                         )
